2020 day 6 C#

+++ 2020/csharp/Day05/Solution.cs
@@ -18,7 +18,7 @@ namespace AdventOfCode2020.Day05
 
         public int PartOne() => _seats.Max();
 
-        public long PartTwo()
+        public int PartTwo()
         {
             var missing =
                 from seat in AllPossibleSeats()
